pyproject.toml
[project]
name = "kenboard"
version = "0.1.20"
description = "Project management dashboard with kanban boards, burndown charts, and category-based organization. Flask API backend with MySQL storage and static HTML generation."
authors = [
    { name = "Q" }
]
requires-python = ">=3.11"
license = { text = "MIT" }
readme = "README.md"

keywords = ["kenboard", "kanban", "project-management", "flask", "mysql", "dashboard"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Framework :: Flask",
    "Topic :: Office/Business :: Scheduling",
]

dependencies = [
    "flask>=3.1",
    "flask-cors>=5.0",
    "flask-login>=0.6",
    "pymysql[rsa]>=1.1",
    "aiosql>=12.0",
    "yoyo-migrations>=8.2",
    "pydantic>=2.11",
    "structlog>=24.0",
    "click>=8.1",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "argon2-cffi>=25.1.0",
    "flask-limiter>=4.1.1",
]

[project.scripts]
kenboard = "dashboard.cli:cli"
ken = "dashboard.ken:cli"

[dependency-groups]
dev = [
    # Package management
    "pdm>=2.26.0",
    "pdm-bump>=0.9.0",
    # Testing
    "pytest>=8.4",
    "pytest-cov>=7.0",
    "pytest-asyncio>=1.2.0",
    "pytest-playwright>=0.6",
    "httpx>=0.24.0",
    # Formatting
    "black>=25.0",
    "isort>=6.0",
    "docformatter>=1.7",
    # Linting
    "ruff>=0.13.0",
    "flake8>=7.3",
    "flake8-docstrings>=1.7.0",
    "flake8-docstrings-complete>=1.3.0",
    # Type checking
    "mypy>=1.18",
    "types-PyYAML>=6.0",
    "types-Flask>=1.1",
    # Quality
    "interrogate>=1.7.0",
    "vulture>=2.14",
    "refurb>=2.2.0",
    "absolufy-imports>=0.3.1",
]
publish = [
    "twine>=6.2.0",
    "keyring>=25.7.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm-bump]
version-files = ["src/dashboard/__init__.py:__version__"]

# -- PDM Scripts --------------------------------------------------------------

[tool.pdm.scripts]

# Testing
test = "pytest tests/ -q --tb=short --ignore=tests/e2e"
test-quick = "pytest tests/ -q --tb=short -x --ignore=tests/e2e"
test-ci = "pytest tests/ -q --tb=short --ignore=tests/integration --ignore=tests/e2e"
test-unit = "pytest tests/unit/ -q --tb=short"
test-integration = "pytest tests/integration/ -q --tb=short"
test-e2e = "pytest tests/e2e/ -q --tb=short"
test-cov = "pytest tests/ -q --tb=short --cov=src/dashboard --cov-report=term-missing --cov-report=html"

# Formatting
format = "black src/ tests/"
isort = "isort src/ tests/"
docformatter = "docformatter --recursive --in-place --wrap-summaries 88 --wrap-descriptions 88 --close-quotes-on-newline src/ tests/"
absolufy = { shell = "find src/ -name '*.py' -exec absolufy-imports {} +" }

# Linting & type checking
lint = "ruff check --fix src/ tests/"
flake8 = "flake8 src/"
typecheck = "mypy src/"
interrogate = "interrogate src/"
vulture = "vulture src/ tests/ vulture_whitelist.py"
refurb = "refurb src/"

# Composite checks
check = { composite = ["isort", "format", "docformatter", "typecheck", "flake8", "interrogate", "refurb", "lint", "vulture", "test-quick"] }

# Build & generation
build-html = "python build.py"
screenshots = "python scripts/screenshots.py"
clean = "rm -rf dist/ build/ .mypy_cache/ .pytest_cache/ htmlcov/ .coverage"

# Install
install = "pdm install"
install-dev = "pdm install -G dev"

# Version
version-show = "python -c \"import dashboard; print(dashboard.__version__)\"" # package still named dashboard internally
version-patch = "pdm bump -v patch"
version-minor = "pdm bump -v minor"
version-major = "pdm bump -v major"

# -- Black --------------------------------------------------------------------

[tool.black]
line-length = 88
target-version = ["py311"]

# -- isort --------------------------------------------------------------------

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
line_length = 88
multi_line_output = 3
include_trailing_comma = true

# -- mypy ---------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = true

# -- pytest -------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "unit: marks unit tests",
    "e2e: marks end-to-end browser tests",
]

# -- coverage -----------------------------------------------------------------

[tool.coverage.run]
source = ["src/dashboard"]
omit = ["tests/*", "src/dashboard/__main__.py"]

[tool.coverage.report]
fail_under = 75
exclude_also = [
    "def __repr__",
    "if __name__ == .__main__.:",
    "pragma: no cover",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
]

# -- vulture ------------------------------------------------------------------

[tool.vulture]
min_confidence = 80
paths = ["src", "tests"]

# -- interrogate --------------------------------------------------------------

[tool.interrogate]
fail-under = 95
ignore-init-method = true
ignore-init-module = true
exclude = ["tests", "docs", "build", "dist"]
generate-badge = "."

# -- docformatter -------------------------------------------------------------

[tool.docformatter]
recursive = true
wrap-summaries = 88
wrap-descriptions = 88
close-quotes-on-newline = true
